+++ CHANGELOG.md
@@ -9,11 +9,16 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 <!-- Please document PR changes here. -->
 
+### Added
+
+* Hooked up the new configurable [polymer-linter](https://github.com/Polymer/polymer-linter) into the editor service. Use polymer.json to configure it.
+
 ### Fixed
 
 * Handle paths more consistently and more correctly, especially paths containing characters which must be encoded when in a URL (e.g. spaces, and all paths on Windows).
   * Handle windows-style file:// urls as well. This should finally bring full support to windows!
 * Updated dependencies.
+
 
 ## 1.2.0 - 2017-01-13